tools/doc-gen/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-gen"
version = "0.1.0"
description = "Simplified documentation generation tool using Claude"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
]

[project.scripts]
doc-gen = "doc_gen.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
